Switch From-Header in emails
Nowadays this can cause problems with spam filters if the domain does not match
the domain of the host the mail is sent from. Although the Internet Message
Format RFC says differently...

https://www.rfc-editor.org/rfc/rfc5322#section-3.6.2

Also we seem to fail to setup good DKIM signatures in the SUSE MTAs for this too.
Let's follow what others do and mangle the From header...