[backend] speed up container publishing with noclean
We only need to query the registry about the tags we know, as we
do not touch all the other ones anyway.

This also needs a change in bs_regpush as we need to extend the tag
list with the cosign tags. This can be enabled by adding '--cosign'
to the ps_regpush query call.