Refactor controller specs
Change from `subject!/it is_expected` to `before/expect(response)?`:

A spec over a controller action has an HTTP request just before the
checks. Most of the checks over a controller action use response as the
implicit subject. Using subject only in the cases where it is really
needed and response in the rest of the cases helps to understand better
this pattern used in controller specs.

See this pattern in the documentation:

https://rspec.info/features/6-1/rspec-rails/controller-specs/controller-spec/