Use logarithmic scale for build result summary chart
Currently the click target is really small on the bars in the chart
when there is a big margin between the values of the individual results
(e.g. 100 Succeding Builds and 1 Failed Build).
Using a logarithmic scale on the y-axis balances the difference
on the bars in the chart.

In order to stay consistent with the ordering of the individual
build results types and to distinguish the click target a little
more I disabled stacking of the bars.