Rename CannedResponse#decision_kind to favored
Replace `favor` with `favored` for consistency and compatibility with
the upcoming refactoring on decisions.

Moreover, the decision form allows canned responses for favored and
cleared decisions. But also for normal comments (decision_kind nil). For
those normal comments the form is now selecting `cleared` which is the
value set in database by default anyway.

Co-authored-by: Eduardo Navarro <enavarro@suse.com>