Do not call `link_to_reportable` if reportable does not exist
We call the `link_to_reportables` helper on the report show view.
This throws an `nil` exception when the reportable is deleted,
since the `reportable_type` gets nullified as well.
`link_to_reportable` works in on the email notification since the
`reportable_type` is still available from the event payload, but
we cannot retrieve it from the report record.

Fixes #16062