Simplify involved_projects/involved_packages query
With those two methods we are checking for a maintainer relationship with
Project/Packages for the user.

The Project/Package default scope hides Project that are access disabled and
where the user has *no* maintainer relationship to.

So both *can not* be true are the same time. No need to skip hidden projects in
the query.