Reduce memory footprint of the Job
By not storing the full BinaryRelease object in the `old_binary_releases` hash
but only the data needed to detect if the BinaryRelease was modified.
In the case it's modified we need to look it up in the database then, but this
should not happen often.

By only iterating through the current and unchanged BinaryRelease of the Repository.

Co-authored-by: Adrian Schröter <adrian@suse.de>