Handle validation errors comming from workflow run
We can simply handle the active record exceptions and render the
appropriate error. Also, we can authorize first the token, there is no
need to even initialize a workflow run if a user is not permitted to run
a workflow.

Also get rid of the create_workflow_run method. The workflow creation is
moved to the 'create' method but the request_headers extraction is moved
to a specific method (to not increase complexity).