Prevent from performing N+1 queries on user involved packages
To show the involved items of a user, results are presented sorted by
project name and project name/package name. To retrieve the project name
of a package, an additional query is performed for every package. This
happens for every involved package, independent of being shown in the
current paginated page or not, in the sorting method.

Include the projects table when querying for involved packages.
Therefore, prevent from performing N+1 queries in the case of user
involved package.