Prevent from erroring for notifications of deleted workflow runs
A failed workflow run event can create a notification. In case this
workflow run is deleted (for example, when its token is removed) an
error is thrown when trying to render the notification.

Check for the existence of the workflow run, so we make sure we can
create a link used to render the notification. If the workflow run
doesn't exist, just render a text (without a link).