[backend] workaround productcomposer naming scheme breaking re-releases of products
The productcomposer uses no suffix for the main medium, but '-Source' and '-Debug'
for the other media types. While we can easily match the latter there is no
way to handle the former in all cases.

Check for '-Source' or '-Debug' as a workaround for now both in the "setrelease"
command and in the "sleimage" publisher code.