Add hidden submit field to request index filter form
Right now its not possible to submit the search input field.
For all other inputs of the same form we have auto submit implemented through JS.
But this makes little sense for the search input.
By adding a hidden submit button we can default to common browser behavior of
submitting the form when pressing the enter button on the selected input.