Special-case SLFO maintenance patchinfos when displaying staging requests
These patchinfos are named "patchinfo.ABC.XYZ" where "ABC" and "XYZ" are
sequences of integers, and this naming scheme is considered an
implementation detail (to avoid collisions when several maintenance
coordinators are working on updates targeting the same codestream).

The numbers are irrelevant for display, the only relevant information is
that these packages are patchinfos, so this change rewrites the package name
to just "patchinfo" in such cases.