[backend] reimplement patching of kiwi product files
We now use Build::SimpleXML to parser/create the kiwi xml file. This
has the advantage that we can record the element offsets and just
patch the parts that need to be changed, leaving all xml comments
intact.

This is needed as the build tool reads the comments to get
the "milestone" data.

This commit also changes the detection of the "rpmhdrs_only" mode
so that we can get rid of the BSKiwiXML module.