Make package labels searchable in package datatable
In order to correctly eager load the label association I had to slighly
change how package_kind is handled. ajax-datatables-rails expects the
association being included in the `get_raw_records` method in order to
apply to apply search and filter values on them.
The `select` and `group` in the db query done in the `data` method afterwards
caused the included association being wrongly handled (previously loaded
labels were incomplete due to the grouping of the records).