Fix capybara session errors
When running feature specs sometimes chromium crashes. One of the
several errors that pop up is this one:

  1.3) Failure/Error: Unable to infer file and line number from
backtrace

    Selenium::WebDriver::Error::InvalidSessionIdError:
      invalid session id

Disabling the use of shared memory seems to fix this error because
Chrome does not use /dev/shm which can be tight on memory on CI
Pipelines or containers, and uses files in /tmp which can be a bit more
forgiving in tight environments at the expense of some performance.