Pin base64 to version 0.1.1
We cannot use anything that uses base64 > 0.1.1 as it is provided by
the system ruby 3.1.

This limits the version of selenium-webdriver that we can use (currently
4.16.0).

Take into account binaries like selenium-manager during build

selenium-manager lives inside bin/linux, not bin, so the build script
crashed in the final steps.

This fix skips subdirectories.