Rework image handling for mkosi, appimage et al
Those build types were useing the default package handler, which is not
a good fit. We do not want to ignore dependencies or use the srcmd5
instead of the hdrmd5 in the meta file.

This commit adds a new generic handler called 'Image', which does the
right thing for image builds.