Show available creator filter options after filter is applied
Before after filtering by a creator, the rest of the available
creators disappeared from the filter dropdown and only the selected
one's remained in the list.
This happend because the list of creators in the filter dropdown
was basedon the requests where the creator filter was already applied.
We have to set the available creators before applying the filter it
self in order to keep the list complete.