[backend]: support multi-stage signing
When building UKIs (unified kernel images) with mkosi we need
to do multi-stage signing. First the image is built from packages,
with the kernel/initrd/etc being put together. The TPM2 policy
digests are calculated, and need to be signed.
After the digests signatures are retrived after the first signing
stage, they are inserted in the UKI, which itself needs to be
signed for secure boot, hence the second consecutive signing stage.

In order to avoid infinite loops, keep a count of consecutive
signing stages in the info file, and stop after 10 no matter what.