[backend] bs_publish: do not use arch-specific directory for mkosi artifacts
mkosi DDIs (.raw) and recipe are already published in the main repo directory.
Publishing in a per-arch subdirectory breaks 'Repotype: checksumsfile' which
is needed to publish a SHA256SUMS signed file, so that systemd-sysupdate can
download images from OBS.

mkosi artifacts can be created with the architecture in the filename with a
simple option in the recipe anyway, so there's no clash if that is used.