Add a rake task to backfill source fields on BsRequestActions
This duplicates the BackfillSourcesOnBsRequestActions data migration which
we have had problems running in our production environment for OBS. It
exited consistently with

```
Mysql2::Error::ConnectionError: Lost connection to MySQL server during query
```

We are suspecting ha-proxy timeouts because of ActiveRecord::Migration
transactions that are running longer than 1h. Doing this in a rake task
removes ActiveRecord::Migration from the equation.