Event listener hooked to the document instead
Using `$(document).on('change', selector, function...` over
`$(selector).change...`: the latter attach an event listener to the
existing matching selectors and only to the exsisting ones at the time
the script runs the first time; the first option instead attaches an
event listener to the `document` who will match the elements living
in the whole document for the entire document life.
The main difference plays when, for instance, an ajax request changes
the content of the document, maybe adding one element that matches
the selector: to let this new element have the event listener
a call back to reimplement the `$(selector).change` should be fired. In
case the event listener was applied with `$(document).on('change'...`
it is the `document` who is holding the event and reacting to that,
therefore the new element will be handled by default with the event
properly without any additional callback.