Improve repositories feature specs
If the test goes too fast the click has not yet reflected in
the following navigated page with the distributions list page, that ends
up being empty. Expecting for the click result and waiting for the ajax
calls make the test more consistent and it reflects the real life
workflow.