Fix error for enum not backed by database column
Since Rails Version 7.1 an exception is raised when defining
an enum that is not backed by the database column.

For reference see:
https://github.com/rails/rails/issues/45668
and
https://github.com/rails/rails/pull/45734

In order to still do this, we can explicitly define the type
via `attribute` on the models.

For reference see:
https://github.com/rails/rails/pull/49769

This is currently blocking the `migration_tests` in our CI.