Split response type: don't compute set_pulse data when not used
`show` route was initially rendering the view only (which behaves as a
placeholder, it does not render anything), and after that
it sends a request (ajax this time) still pointing to `show`
route, which reloads the entire data and then it renders them. This
makes the `set_pulse` data to be computed twice, the first time for the
show template, and the second time to really get and present the pulse
data.

The current change splits the show response type to behave differently:
- `show` just to render the view (show.html.haml), no unused data computed
- `show`+`set_pulse` async load pulse data and render the rest (show.js.erb)