Load Rails 7.0 defaults
Get rid of the following deprecation warning:

DEPRECATION WARNING: Support for
`config.active_support.cache_format_version = 6.1` has been deprecated
and will be removed in Rails 7.2.

Check the Rails upgrade guide at
https://guides.rubyonrails.org/upgrading_ruby_on_rails.html#new-activesupport-cache-serialization-format
for more information on how to upgrade.

This is a step needed to be done previous to the upgrade of Rails to
version 7.2.