Fix: html id is meant to be unique in the page
Both `add_repository_path...` and `add_repository_from_project...`
reuse the same partial that implements the `autocomplete` functionality
targetting the html element by the `id` attribute.
If both input fields has the same `id` the autocomplete crashes on the
first one and works only on the last one that was implemented in the
page.