Get rid of stale and useless ajax parameters
See https://api.jquery.com/Jquery.ajax/

`type` is an alias for `method` used for jQuery < 1.9, we're much ahead
of it, and it is a `get` request, the default method, no reason to
specify it can be omit.

`dataType` defines the expected response type of data, but the framework
is smart enough to detect it based on the response it receives.