[api] allow to process certain operations on scmsync operations
Allowing esp. to remove projects or alter meta data via request
when they are scmsync managed. This is fixing esp. the crashes
in delayed jobs for the cleanup jobs.

Should be done via different policy rules instead makeing a difference
between meta and source modification of a project, but the merge
request 15476 providing this is not merged yet.