Remove override of $link-hover-color variable
Previously, the color of the links changed from $blue to $green when
hovering on them. Reproducing this behaviour in different color modes is
not an easy task.

Removing the value assigned to the `$link-hover-color` variable, allows
Bootstrap link utilities to define a color for hovering on links
automatically.

See as reference:

https://getbootstrap.com/docs/5.3/helpers/colored-links/#link-utilities