Support readonly Package in policy
If the Package is readonly it means it is coming from a
remote project link, scmsync project or is a special
Package like `_project`.

Some things users can do for this packages if they can write the project.

Also switch away from `user.can_modify?` it just calls the other two.