Fix checking read access of deleted project/packages
Project.get_by_name(project) raises if the project is deleted which
made it impossible to check for a package in a deleted project.

Also there is *no* significance to the revisions. Either you can fetch meta
(the last revision) or the project/package never existed. I guess I know where
this came from, fetching revisions of a project that never existed returns an
empty revisions list.

```
[1] pry(main)> Backend::Api::Sources::Project.revisions('rutzelfulz13892')
=> "<revisionlist>\n</revisionlist>\n"
[2] pry(main)> Backend::Api::Sources::Package.revisions('rutzelfulz13892', '38sjwj2')
=> "<revisionlist>\n</revisionlist>\n"
```

Probably confused whoever wrote this first...