Fix reopening a previously staged request
Release managers:

1. Stage the BsRequest into a StagingProject
  - which creates a Review.by_project for StagingProject
  - and adds the BsRequest to StagingProject.staged_requests
2. Find out that the BsRequest is not up to standard
3. Decline the BsRequest
4. Unstage the BsRequest from the StagingProject
  - which removes the BsRequest from StagingProject.staged_requests
  - and declines the Review.by_project for StagingProject

The submitter:

1. disagrees and reopens the BsRequest
   - which sets the Review.by_project for StagingProject to new

Now we have a Review.by_project for StagingProject without the
BsRequest being in StagingProject.staged_requests. That breaks the
staging workflow because you always have the dangling review on the
BsRequest for a StagingProject the request is no in anymore...

If we set the Review.by_project for StagingProject to accepted
when we unstage the BsRequest from the StagingProject then reopening
the BsRequest will not touch the Review.by_project for StagingProject.