Add staging reviews to a declined request without changing the request state
`BsRequest.addreview` and `BsRequest.change_review_state` also change the state of the request.
Avoid this if the `BsRequest` is declined before un-staging because that would require permissions
to reopen and decline the `BsRequest`. A Staging manager does not necessarily have those permissions.
Especially if we are using `Bsrequest.approver`.