Revert "Use dropdown instead of popover for double links"
This reverts commit e2388cfb23be39102166b66eb62d1f65298ddff0.
Dropdown is used in this case inside of an accordion, and when the
accordion content has only one/few rows, there is no space to show the
dropdown content, and this makes the accordion content box to overflow
and scroll to see the dropdown content. Using the popover makes this
glitch go away.