Better `where` than object converted to array
In many parts of the code we need to iterate over a bunch of objects,
even if sometimes there is only one or none.

To retrieve those objects we better use `Model.where()` than
`[Model.find_by()].compact` or similar.

The second option is considered an anti-pattern and returns an array that
could includes nil values (that's why we need `.compact` sometimes).
The second will return an ActiveRecord::Relation with zero, one or more
elements but never nil. It also implies a better performance.