Remove unnecesary `include_all_packages` parameter
Previously, when the `include_all_packages` parameter was set to true,
the `Project#get_by_name` method would raise a `ReadAccessError`
exception if there was a package that:

- had the 'sourceaccess' flag,

- belonged to the current project, and

- the current project could not be modified by the current user.

However, the check to determine whether the current project can be
modified by the current user is performed immediately after this code
anyway. Therefore, we can safely remove the `include_all_packages`
parameter and all associated conditional logic.