[backend] share ca-certificates with service container
We are using self signed certificates for our appliances.
Without this patch using the local gitea installation fails unless the
ca-certs storage is not shared with the container because of this local
cert.
This patch adds the directory `/var/lib/ca-certificates` to the list of
shared volumes if the called service is obs_scm_bridge.