[backend] bs_regpush: implement handling for rate limiting
This makes use of the new "toomanyrequests" feature of BSRPC. When
running into rate limiting, bs_regpush will return an exit status
from 10 to 20, which encodes the minutes to wait.

The published will check for an exit status in that range and
use it when calculating the publish retry time.

This commit also makes bs_regpush check for http error codes
that indicate a setup problem and exit with status 2 in that
case. The publisher will detect this, log a critical error
and drop the publish event so that no retry happens.