Fix date of add LimitReleaseSourceProject attribute data migration
Rename the data migration file for adding the LimitReleaseSourceProject
attribute type to a time just before the data migration to add the roles
for this same attribute.

Otherwise, instances which run their data migrations, which previosly
had run data migrations with dates after the previous 20220429123314
timestamp, will thrown an exception on the add roles data migration.