Remove session checking
Authorization is not the place to check for Authentication.

The weak version of this method (`check_source_access?`) uses
`User.possibly_nobody`. So only if someone makes the nobody user admin, gives
it a `Role` with global `source_access` permission or adds this `User` to it's
`Project`/`Package` this would hit. If someone so desperately tries to shoot
themselves in the foot, let them.