Rename `require_package` method to `set_package`
In the majority of controllers we use the `set_package` alias
to call `require_package`. Also the controllers that have their
own logic for this method and overwrite the inheritance of the
method use the `set_package` naming. Lets get rid of the
alias and stick to one name for it.