Fix rendering of `BsRequestActionSourceAndTargetComponent`
In the previous update of the ViewComponent Gem
(https://github.com/openSUSE/open-build-service/pull/18269), the
`capture` block around the view code in the `combine`
method was dropped. This led to only the last line (last view
tag) being returned and therfore only the link to the target
of the bs_reques_action being rendered.
Simply reintroducing the use of the `capture` helper doesn't workout
since the usage of it depends on the view context that only exists
once a ViewComponent is passed to the Rails render pipeline.

A little refactor of moving the view code to a view template
and handling some of the logic previously handled outside of
the view component is the easiest way of fixing it.

Fixes https://github.com/openSUSE/open-build-service/issues/18302