Disable diff request for files with no diff
Sometimes we don’t receive a diff from the backend for certain files.
In these cases, the diff content is just the number of lines changed without the
actual diff, and sometimes the diff is not available at all. This does not depend
on the state of the file. We were checking this based on the file’s state, but the
file could be in any state (added, deleted, changed, or renamed) and still have no diff.