[backend] bs_service: implement "noprefix" mode on the server
The old noprefix implementation stripped the service: prefix after
receiving the files from the service server. This does not work
correctly if a generated file contains a ':' character. It's not
clear if the received file 'service:foo:bar:baz' is 'bar:baz'
generated by service foo or 'baz' generated by services foo and
bar.

So add an option to skip the prefix generation and make use
of it for obsscm service runs.