Fix turbo frame id for package multibuild flavors
Right now one of the turbo frames adds the multibuild flavor
to the html id, the other one not. This leads to turbo not
being able to locate the frame and rendering the partial
outside of the frame.
The `package_name` variable used on one of the frames contains
the multibuild flavor. Lets use the variable on both so they
match in the case of a package with multibuild flavors.

Fixes #18487

Co-authored-by: Saray Cabrera Padrón <scabrerapadron@suse.de>