[api] Fix usage of scmsync projects via project links
It is not enough to check only the project itself for scmsync
otherwise permissions are checked on package A and execution is
done on package B.

Broke with 02afcf6faa2 for scmsync which made this a special thing for
scmsync, but it needs to work anyway in general as it does for
remote packages. It means we would also need to duplicate all
permission testing now.

issue#18205