Persist comment text after page switch or refresh
To prevent the loss of draft comments when switching between
pages (especially the tabs of the request workflow) we need to
store the text in the session store of the browser and reapply
it on page load at the corresponding form.

With this commit this functionality is given for comments on
request conversations, request inline diffs, packages and projects.
It works for editing or replying as well.

For now lets use the session store over the local store since
it gives a little more security and requires less attention in
terms of cleaning up the store. The stored comments will only
be available inside the current browser session and be gone
after closing the browser window/tab.

Fixes #18504