Let the request index item not be a single blob link
The list item was a single blob link, but it contains
useful other clickable links the user wants to be able to reach.

This change makes the item an heterogeneous object with multiple links
in its own content. The title keeps the role to link to the request
show page.