Do not render build results chart if there are too many `bs_request_actions`
The query produced by this can be too expensive when there are too many
`bs_request_actions` involved. This causes performance issues on the
backend, since passenger instances pile up with those calls waiting to be
processed. As a quick fix, skip rendering of this for now if there are
more than 300 `bs_request_actions`.