Consider the assigner in assignment policies
The policy doesn't check if the assigner is authorized to add
an assignee to a package. As of now everyone can add an assigne.
We have to check if the assigner has the required role on the package or
project. Also, we can drop checking for the assignee here, this is done
through a validation on the model level. No need to do it here as well.