Remove check of `count` variable in notifications controller spec
We already test if the corresponding notification got updated or
not. The value of `count` is just the return value of the
`update_all` call. Since the variable is a local variable now,
we can not test it this way anymore, but it is not required to
do so anyway. The main logic if the corresponding notification
got updated or not is covered, so we can drop it...