Pin `strscan` bundled gem version to the one used by ruby3.4
"strscan" is one of the default/standard gems shipped with ruby.
We should pin the version used in the bundle to the one coming
from the ruby interpreter currently used in production (ruby 3.4.7)
to prevent accidental updates of it (which leads to version mismatch
errors, preventing phusion passenger from spawning).