Fix empty query parameters in `expand_params` method
Previously, there was no distinction between parameters passed to the
backend with an empty value.

For example, when parsing the query string `?expand`, Rails would
include the key but with a nil value in the `params` hash. When this
hash was later serialized back into a query string using the `.to_query`
method, it erroneously resulted in `?expand=`.

This incorrect modification caused the backend to throw a 400 error.

This commit resolves the issue by splitting query parameters into two
distinct groups of parameters and handling them differently.