[backend] Avoid duplicate updateinfo_id when not defined
We did set the updateinfo id to the project name in case it
was not defined elsewhere. This leads to conflicts when multiple
patchinfos were inside the project.

We extend it now with the package name to avoid duplicates.

This is an incompatible change, but all production configurations
define the patchinfo id. Also it would lead to not usable repository
meta data, so it is unlikely a regression.